requirements.txt
git+https://github.com/orthose/api-airfranceklm-python-sdk.git
nltk
discord
python-dotenv
pandas
parsedatetime
spacy
text2num